FluxSegmentationModels

Documentation for FluxSegmentationModels.

FluxSegmentationModels.EncoderConfig

FluxSegmentationModels.Conv

FluxSegmentationModels.ConvBlock

FluxSegmentationModels.FPNDecoder

FluxSegmentationModels.SegFormerDecoder

FluxSegmentationModels.SeparableConv

FluxSegmentationModels.UNetDecoder

FluxSegmentationModels.Upsample

FluxSegmentationModels.build_encoder

FluxSegmentationModels.encoder_features

FluxSegmentationModels.EncoderConfig — Type

EncoderConfig

Super type of all encoder models. The type parameter F denotes the number of features produced by each block of the encoder.

Example Implementation

struct ResNet <: EncoderConfig
    depth::Int
    pretrain::Bool
end

function ResNet(;depth=50, pretrain=true)
    @argcheck depth in (18,34,50,101,152)
    return ResNet(depth, pretrain)
end

function encoder_features(e::ResNet)
    @match e.depth begin
        18 || 34 => [EncoderFeature(64,4), EncoderFeature(128,8), EncoderFeature(256,16), EncoderFeature(512,32)]
        50 || 101 || 152 => [EncoderFeature(256,4), EncoderFeature(512,8), EncoderFeature(1024,16), EncoderFeature(2048,32)]
    end
end

function build_encoder(e::ResNet; inchannels=3)
    resnet = _build_resnet(e.depth, inchannels, e.pretrain)
    Flux.Chain(
        Flux.Chain(resnet[1]..., resnet[2]...),
        resnet[3],
        resnet[4],
        resnet[5],
    )
end

function _build_resnet(depth::Int, inchannels::Int, pretrain::Bool)
    return Metalhead.ResNet(depth; inchannels, pretrain).layers[1]
end

FluxSegmentationModels.Conv — Method

Conv(kernel_size::Tuple{Int,Int}, in::Int, out::Int; act=Flux.relu, norm=:BN, groups=1, dilation=1)

A block of convolutional layers with optional batch normalization.

Parameters

kernel_size: The size of the filter to use for each layer.

in: The number of input features.

out: The number of output features.

act: The activation function to apply after each convolution.

norm: Specifies the type of layer normalization. One of :BN, :LN, or nothing.

group: The number of groups to divide the convolution into.

dilation: The dilation rate for the kernel.

FluxSegmentationModels.ConvBlock — Method

ConvBlock(kernel_size::Tuple{Int,Int}, in::Int, out::Int; act=Flux.relu, depth=2, norm=:BN)

A block of convolutional layers with optional batch normalization.

Parameters

kernel_size: The size of the filter to use for each layer.

in: The number of input features.

out: The number of output features.

act: The activation function to apply after each convolution.

depth: The number of successive convolutions in the block.

norm: Specifies the type of layer normalization. One of :BN, :LN, or nothing

FluxSegmentationModels.FPNDecoder — Method

FPNDecoder(encoder_dims::AbstractVector{<:Integer}, encoder_scales::AbstractVector{<:Integer}; embed_dim=256, dropout=0.0, batch_norm=true)

A MLP-style decoder as used in SegFormer. Expects a tuple of block-wise activations as input.

Parameters

encoder_dims: The feature dimension of each encoder block output ordered from first to last.

encoder_scales: The scale of each encoder block output with respect to the input ordered from first to last.

embed_dim: The size of the embedding dimension to use. Shared by all layers.

dropout: The dropout probability to use after the last layer.

batch_norm: Use batch normalization after each convolution.

FluxSegmentationModels.SegFormerDecoder — Method

SegFormerDecoder(feature_dims::Vector{Int}, feature_scales::Vector{Int}; embed_dim=768, dropout=0.0, norm=:BN)

A MLP-style decoder as used in SegFormer. Expects a tuple of block-wise activations as input.

Parameters

encoder_dims: The feature dimension of each encoder block output ordered from first to last.

encoder_scales: The scale of each encoder block output with respect to the input ordered from first to last.

embed_dim: The size of the embedding dimension to use. Shared by all layers.

dropout: The dropout probability to use after the last layer.

norm: One of either :BN or :LN.

FluxSegmentationModels.SeparableConv — Method

SeparableConv(kernel::Tuple{Int,Int}, in::Int, out::Int; stride=1, act=identity, pad=Flux.SamePad(), norm=:BN)

A block of convolutional layers with optional batch normalization.

Parameters

kernel_size: The size of the filter to use for each layer.

in: The number of input features.

out: The number of output features.

stride: The stride of the convolution operation.

act: The activation function to apply after each convolution.

norm: Specifies the type of layer normalization. One of :BN, :LN, or nothing.

FluxSegmentationModels.UNetDecoder — Method

UNetDecoder(encoder_dims, encoder_scales; decoder_dims=[64,128,256,512], batch_norm=true)

Construct a U-Net style decoder that expects a tuple of WHCN arrays corresponding to the output of each block of a matching encoder.

Parameters

encoder_dims: The feature dimension of each encoder block output ordered from first to last.

encoder_scales: The scale of each encoder block output with respect to the input ordered from first to last.

decoder_dims: The feature dimension of each decoder block ordered from top to bottom.

batch_norm: If true, a batch norm operation will be applied after each convolution in the decoder.

FluxSegmentationModels.Upsample — Method

Upsample(factor::Int; method=:nearest)

Construct a layer that takes an image with shape WHCN and upsamples it by factor.

Parameters

factor: The factor by which to upscale the input image.

method: The type of resampling to use; must be one of :nearest or :bilinear.

FluxSegmentationModels.build_encoder — Function

build_encoder(encoder::EncoderConfig)

Constructs an encoder model based on the provided EncoderConfig configuration.

Parameters

encoder: An EncoderConfig object specifying the architecture and configuration of the encoder to be built.

Returns

A standard Flux.Chain layer containing each block of the encoder. The returned encoder is ready to be integrated into more complex architectures like U-Net or used as a standalone feature extractor.

Example

encoder = build_encoder(ResNet(depth=50, pretrain=true))

FluxSegmentationModels.encoder_features — Function

encoder_features(encoder::EncoderConfig)

Returns the dimension and scale of the provided encoder.